% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-combine.R
\name{combine_docs}
\alias{combine_docs}
\title{Combine specific documents available in the covidphtext package into a single
dataset}
\usage{
combine_docs(docs = c("resolution", "press release"))
}
\arguments{
\item{docs}{A vector of terms to search for in the document names}
}
\value{
A tibble of all document types called for.
}
\description{
Combine specific documents available in the covidphtext package into a single
dataset
}
\examples{
combine_docs(docs = "resolution")

}
