% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidphtext.R
\docType{package}
\name{covidphtext}
\alias{covidphtext}
\alias{covidphtext-package}
\title{Utilities to Extract Text Data From COVID-19-Related Resolutions and Policies
from the Philippines}
\description{
To assess possible impact of various COVID-19 prediction models on
Philippine government response, text from various resolutions, press
releases and legislation has been collected using data mining approaches
implemented in R. This package includes functions used for this data
mining process and datasets of text that have been collected and processed
for use in text analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/como-ph/covidphtext}
  \item \url{https://como-ph.github.io/covidphtext}
  \item Report bugs at \url{https://github.com/como-ph/covidphtext/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

Authors:
\itemize{
  \item Chris Mercado \email{cegmercado@gmail.com} (\href{https://orcid.org/0000-0002-5588-9543}{ORCID})
  \item Robert Medina \email{jcmedina1@up.edu.ph} (\href{https://orcid.org/0000-0001-6647-5453}{ORCID})
}

}
\keyword{internal}
