% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-get_iatf.R
\name{get_iatf_gazette}
\alias{get_iatf_gazette}
\title{Get list of links to IATF resolutions from Philippines Officie Gazette}
\usage{
get_iatf_gazette(base, pages = 1:5)
}
\arguments{
\item{base}{URL to the IATF resolutions webpage in the Official Gazette.
This is currently at \url{https://www.officialgazette.gov.ph/section/laws/other-issuances/inter-agency-task-force-for-the-management-of-emerging-infectious-diseases-resolutions/}}

\item{pages}{A vector of paginated webpages in which the IATF resolutions
are available. This currently defaults to 1:5 as there are currenlty 5
pages.}
}
\value{
A tibble containing absolute links to all the current IATF
  resolutions at time of extraction from the Official Gazette. The tibble
  contains the following information:

\itemize{
  \item \code{id}{Resolution number}
  \item \code{title}{Title of resolution}
  \item \code{date}{Date (in <YYYY-MM-DD> format) resolution was issued}
  \item \code{source}{Source of resolution. This is by default from IATF}
  \item \code{type}{Type of document. This is by default a resolution}
  \item \code{url}{Absolute URL for PDF of resolution}
  \item \code{checked}{Date (in <YYYY-MM-DD format) table was extracted. This
    is by default provided by \code{Sys.Date()}}
}
}
\description{
Get list of links to IATF resolutions from Philippines Officie Gazette
}
\examples{
base <- "https://www.officialgazette.gov.ph/section/laws/other-issuances"
agency <- "inter-agency-task-force-for-the-management-of-emerging-infectious-diseases-resolutions/"

get_iatf_gazette(base = paste(base, agency, sep = "/"), pages = 1)

}
