% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iatf.R
\name{get_iatf_links}
\alias{get_iatf_links}
\title{Get list of links to IATF resolutions from Department of Health website}
\usage{
get_iatf_links(base = base_urls[[1]])
}
\arguments{
\item{base}{URL to the IATF resolutions webpage. This is currently at
\url{http://www.doh.gov.ph/COVID-19/IATF-Resolutions/}}
}
\value{
A tibble containing absolute links to all the current IATF
  resolutions at time of extraction. The tibble contains the following
  information:

\itemize{
  \item \code{id}{Resolution number}
  \item \code{title}{Title of resolution}
  \item \code{date}{Date (in <YYYY-MM-DD> format) resolution was issued}
  \item \code{source}{Source of resolution. This is by default from IATF}
  \item \code{type}{Type of document. This is by default a resolution}
  \item \code{url}{Absolute URL for PDF of resolution}
  \item \code{checked}{Date (in <YYYY-MM-DD format) table was extracted. This
    is by default provided by \link{Sys.Date}}
}
}
\description{
Get list of links to IATF resolutions from Department of Health website
}
\examples{
get_iatf_links()

}
