% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iatf.R
\name{get_iatf_page}
\alias{get_iatf_page}
\title{Get contents of an IATF page from the Official Gazette}
\usage{
get_iatf_page(page)
}
\arguments{
\item{page}{A URL of a page of the IATF Resolutions in the Official Gazette
website.}
}
\value{
A tibble containing absolute links to all the current IATF
  resolutions from specified page at time of extraction from the Official
  Gazette. The tibble contains the following information:

\itemize{
  \item \code{id}{Resolution number}
  \item \code{title}{Title of resolution}
  \item \code{date}{Date (in <YYYY-MM-DD> format) resolution was issued}
  \item \code{source}{Source of resolution. This is by default from IATF}
  \item \code{type}{Type of document. This is by default a resolution}
  \item \code{url}{Absolute URL for the webpage of the resolution}
  \item \code{checked}{Date (in <YYYY-MM-DD format) table was extracted. This
    is by default provided by \code{Sys.Date()}}
}
}
\description{
Get contents of an IATF page from the Official Gazette
}
\examples{
pages <- list_iatf_pages(pages = 1)
get_iatf_pages(pages = pages)

}
