% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iatf.R
\name{get_iatf_pages}
\alias{get_iatf_pages}
\title{Get contents of the IATF pages from the Official Gazette}
\usage{
get_iatf_pages(pages)
}
\arguments{
\item{pages}{A character vector of URLs of all pages of the IATF Resolutions
in the Official Gazette. This can be created using \code{list_iatf_pages}}
}
\value{
A tibble containing absolute links to all the current IATF
  resolutions at time of extraction from the Official Gazette. The tibble
  contains the following information:

\itemize{
  \item \code{id}{Resolution number}
  \item \code{title}{Title of resolution}
  \item \code{date}{Date (in <YYYY-MM-DD> format) resolution was issued}
  \item \code{source}{Source of resolution. This is by default from IATF}
  \item \code{type}{Type of document. This is by default a resolution}
  \item \code{url}{Absolute URL for the webpage of the resolution}
  \item \code{checked}{Date (in <YYYY-MM-DD format) table was extracted. This
    is by default provided by \code{Sys.Date()}}
}
}
\description{
Get contents of the IATF pages from the Official Gazette
}
\examples{
base <- "https://www.officialgazette.gov.ph/section/laws/other-issuances"
agency <- "inter-agency-task-force-for-the-management-of-emerging-infectious-diseases-resolutions/"
pages <- list_iatf_pages(base = paste(base, agency, sep = "/"), pages = 1)
get_iatf_pages(pages = pages)

}
