% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iatf.R
\name{get_iatf_pdf}
\alias{get_iatf_pdf}
\title{Get a specific IATF resolution PDF file}
\usage{
get_iatf_pdf(links, id)
}
\arguments{
\item{links}{A links/URLs tibble produced by either \code{get_iatf_links()}
or \code{get_iatf_gazette()} functions}

\item{id}{A vector of number/s of IATF resolution/s to be retrieved}
}
\value{
A temporary directory path or a vector of temporary directory paths
  pointing to a temporary file/s for PDF of IATF resolution/s required
}
\description{
Get a specific IATF resolution PDF file
}
\examples{
get_iatf_pdf(links = iatfLinksGazette, id = 29)

}
