% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iatfGuidelineOmnibus}
\alias{iatfGuidelineOmnibus}
\title{COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Guideline Omnibus}
\format{
A tibble with 753 rows and 6 columns:
\describe{
  \item{\code{linenumber}}{Linenumber}
  \item{\code{text}}{Character string of text about 70 characters}
  \item{\code{source}}{Issuer of guideline}
  \item{\code{type}}{Type of document}
  \item{\code{id}}{Text document identifier}
  \item{\code{date}}{Date in <YYYY-MM-DD> format when guideline was issued}
}
}
\source{
The Official Gazette of the Republic of the Philippines
}
\usage{
iatfGuidelineOmnibus
}
\description{
COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Guideline Omnibus
}
\examples{
iatfGuidelineOmnibus

}
\keyword{datasets}
