% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iatfLinks}
\alias{iatfLinks}
\title{IATF resolutions URLs}
\format{
A tibble with 25 rows and 7 columns:
\describe{
  \item{\code{id}}{Unique identifier}
  \item{\code{title}}{Title of resolution}
  \item{\code{date}}{Date resolution released. Format <YYYY-MM-DD>}
  \item{\code{source}}{Issuing agency. IATF by default}
  \item{\code{type}}{Type of issuance. Resolution by default}
  \item{\code{link}}{Absolute URL of IATF resolution}
  \item{\code{checked}}{Date (in <YYYY-MM_DD> format) information was
    extracted. Default is \code{Sys.Date()}}
}
}
\source{
\url{https://www.doh.gov.ph/COVID-19/IATF-Resolutions}. Updated
  as of 19 May 2020.
}
\usage{
iatfLinks
}
\description{
IATF resolutions URLs
}
\examples{
iatfLinks

}
\keyword{datasets}
