% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iatfResolution46}
\alias{iatfResolution46}
\title{COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Resolution Number 46}
\format{
A tibble with 73 rows and 6 columns:
\describe{
  \item{\code{linenumber}}{Linenumber}
  \item{\code{text}}{Character string of text about 70 characters}
  \item{\code{source}}{Issuer of resolution}
  \item{\code{type}}{Type of document}
  \item{\code{id}}{Text document identifier}
  \item{\code{date}}{Date in <YYYY-MM-DD> format when resolution was made}
}
}
\source{
\url{https://www.officialgazette.gov.ph/downloads/2020/06jun/20200615-IATF-RESOLUTION-NO-46.pdf}
}
\usage{
iatfResolution46
}
\description{
COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Resolution Number 46
}
\examples{
iatfResolution46

}
\keyword{datasets}
