% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iatfResolution50A}
\alias{iatfResolution50A}
\title{COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Resolution Number 50A}
\format{
A tibble with 113 rows and 7 columns:
\describe{
  \item{\code{linenumber}}{Linenumber}
  \item{\code{text}}{Character string of text about 70 characters}
  \item{\code{source}}{Issuer of resolution}
  \item{\code{type}}{Type of document}
  \item{\code{id}}{Text document identifier}
  \item{\code{section}}{Resolution section - heading, preamble, operative,
    endorsement, or certification}
  \item{\code{date}}{Date in <YYYY-MM-DD> format when resolution was made}
}
}
\source{
\url{https://www.officialgazette.gov.ph/downloads/2020/06jun/20200629-IATF-RESOLUTION-NO-50A.pdf}
}
\usage{
iatfResolution50A
}
\description{
COVID-19 Inter-Agency Task Force for the Management of Emerging Infectious
Diseases Resolution Number 50A
}
\examples{
iatfResolution50A

}
\keyword{datasets}
