% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-combine.R
\name{combine_iatf}
\alias{combine_iatf}
\title{Combine all Inter-Agency Task Force for the management of emerging
infectious diseases (IATF) resolutions datasets into a single dataset}
\usage{
combine_iatf(docs = "iatfResolution", res = 9:19)
}
\arguments{
\item{docs}{A vector of terms to search for in the document names}

\item{res}{A vector of document identifiers to concatenate. Default is 9:19.}
}
\value{
A tibble of all the IATF resolutions
}
\description{
Combine all Inter-Agency Task Force for the management of emerging
infectious diseases (IATF) resolutions datasets into a single dataset
}
\examples{
combine_iatf()

}
