% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-get_iatf.R
\name{get_iatf_pdf}
\alias{get_iatf_pdf}
\title{Get a specific IATF resolution PDF file}
\usage{
get_iatf_pdf(base = "https://www.doh.gov.ph/COVID-19/IATF-Resolutions", id)
}
\arguments{
\item{base}{URL to the IATF resolutions webpage. This is currently at
\url{https://www.doh.gov.ph/COVID-19/IATF-Resolutions}}

\item{id}{A vector of number/s of IATF resolution/s to be retrieved}
}
\value{
A temporary directory path pointing to a temporary file for PDF
  of IATF resolution required
}
\description{
Get a specific IATF resolution PDF file
}
\examples{
get_iatf_pdf(id = 29)

}
