% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iatf.R
\name{get_iatf_pdf}
\alias{get_iatf_pdf}
\title{Get a specific IATF resolution PDF file given a link}
\usage{
get_iatf_pdf(link)
}
\arguments{
\item{link}{A URL to an IATF resolution PDF}
}
\value{
A temporary directory path for PDF of IATF resolution specified
}
\description{
Get a specific IATF resolution PDF file given a link
}
\examples{
get_iatf_pdf(link = iatfLinksGazette$url[1])

}
